refactor(desktop): split persona request types into types/requests.rs
types.rs sits at 996/1000 under the file-size guard and the quad fields
about to land on CreatePersonaRequest/UpdatePersonaRequest would breach
it. Pure move; the pub use re-export keeps every existing path working.
