+++ desktop/src-tauri/src/managed_agents/types/requests.rs
@@ -1,0 +1,49 @@
+//! Persona command request types, split from `types.rs` (file-size cap).
+
+use std::collections::BTreeMap;
+
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CreatePersonaRequest {
+    pub display_name: String,
+    pub avatar_url: Option<String>,
+    pub system_prompt: String,
+    #[serde(default)]
+    pub runtime: Option<String>,
+    #[serde(default)]
+    pub model: Option<String>,
+    #[serde(default)]
+    pub provider: Option<String>,
+    #[serde(default)]
+    pub name_pool: Vec<String>,
+    /// Environment variables for agents created from this persona.
+    #[serde(default)]
+    pub env_vars: BTreeMap<String, String>,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UpdatePersonaRequest {
+    pub id: String,
+    pub display_name: String,
+    pub avatar_url: Option<String>,
+    pub system_prompt: String,
+    #[serde(default)]
+    pub runtime: Option<String>,
+    #[serde(default)]
+    pub model: Option<String>,
+    #[serde(default)]
+    pub provider: Option<String>,
+    #[serde(default)]
+    pub name_pool: Vec<String>,
+    /// Environment variables for agents created from this persona.
+    ///
+    /// Absent (`None`) = don't touch the stored value (caller didn't include
+    /// the field). `Some(map)` = replace entirely (empty map clears all).
+    /// Defaulting an omitted field to an empty map would silently erase
+    /// stored credentials when an unrelated field is edited.
+    #[serde(default)]
+    pub env_vars: Option<BTreeMap<String, String>>,
+}
